+++ src/crypto.rs
@@ -11,7 +11,7 @@ use crypto_secretbox::{
 };
 use data_encoding::{HEXLOWER, HEXLOWER_PERMISSIVE};
 use rand::Rng;
-use serde::{Serialize, Serializer};
+use serde::{Deserialize, Deserializer, Serialize, Serializer, de};
 use serde_json as json;
 use zeroize::{Zeroize, ZeroizeOnDrop};
 
@@ -101,13 +101,25 @@ fn random_padding_amount() -> u8 {
 
 /// An encrypted message. Contains both the ciphertext and the nonce.
 pub struct EncryptedMessage {
+    /// Ciphertext of the encrypted message
     pub ciphertext: Vec<u8>,
+    /// Nonce of the encrypted message
     pub nonce: Nonce,
 }
 
 /// The public key of a recipient.
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct RecipientKey(pub PublicKey);
+
+impl<'de> Deserialize<'de> for RecipientKey {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+        FromStr::from_str(&s).map_err(de::Error::custom)
+    }
+}
 
 impl From<PublicKey> for RecipientKey {
     /// Create a `RecipientKey` from a `PublicKey` instance.
@@ -231,7 +243,9 @@ pub fn encrypt_file_msg(
 /// returned by [`decrypt_file_data`].
 #[derive(Clone)]
 pub struct FileData {
+    /// The unencrypted data of the file
     pub file: Vec<u8>,
+    /// Optional unencrypted thumbnail, for e.g. a photo or video
     pub thumbnail: Option<Vec<u8>>,
 }
 
@@ -241,7 +255,9 @@ pub struct FileData {
 /// returned by [`encrypt_file_data`].
 #[derive(Clone)]
 pub struct EncryptedFileData {
+    /// The encrypted data of the file
     pub file: Vec<u8>,
+    /// Optional encrypted thumbnail, for e.g. a photo or video
     pub thumbnail: Option<Vec<u8>>,
 }