+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Possible log types:
 
 - [added] Add support for generic public key caching:`lookup_pubkey_with_cache`
   API method and PublicKeyCache` trait (#77)
+- [added] Add support for bulk lookups: `lookup_pubkeys_bulk` and `lookup_ids_bulk` (#89)
+- [added] Add support for bulk sending through `E2eApi::send_bulk` (#89)
+- [changed] Increase default HTTP client timeout from 10s to 120s
 
 ### v0.18.0 (2024-07-13)